move tile cache to its own file
